fix(classified ads list): fix of infinite scroll bug
- fix of the bug by moving the pagin increment logic up to the useJobsAdsListDataHook
